Move network related functions to new `requests.py` module
