+++ substreams_firehose/config/ui/forms/main_config_edit.py
@@ -9,7 +9,7 @@ import logging
 from npyscreen import notify
 from requests.exceptions import RequestException
 
-from substreams_firehose.utils import get_auth_token
+from substreams_firehose.requests import get_auth_token
 from substreams_firehose.config.parser import Config
 from substreams_firehose.config.ui.forms.generic import CategorizedItemDisplayForm
 from substreams_firehose.config.ui.widgets.custom import notify_yes_no
